core/pyproject.toml
[project]
name = "bam-core"
version = "0.0.1"
description = "Shared code for BAM Automations"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyairtable==2.1.0.post1",
    "requests==2.31.0",
    "urllib3==1.26.16",
    "boto3==1.28.34",
    "PyYAML==6.0.1",
    "python-dotenv==1.0.0",
    "phonenumbers==8.13.19",
    "email-validator==2.0.0.post2",
    "charset-normalizer==3.2.0",
]

[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools>=68.1.2", "wheel>=0.41.2"]
build-backend = "setuptools.build_meta"
